Dev to Main Merge - Forget Password and Search on Mobile Functionality
* billing rate changes - phase 1

* fix saving company

* clear drop down fixes

* company edit fix

* new droplet changes

* Mobile App Changes

* search functionality in mobile

* allowing cordinator and admin to talk to client

* Implement clientId query parameter handling in RealTimeChat for auto-selecting users

* build error fix

* bug fixes for via sr to chat

* bug fixes

* verbiage change

* verbiage change

* first commit

* fix 2 - imp

* phase 1 - password reset

* digital occean fix for email

* using mailgun now

* env changes

* mailgun fix

* email config changes

* disclaimer text

* mobile related changes

* Verbiage Changes

+++ SM_MentalHealthApp.Shared/InvoicingModels.cs
@@ -36,10 +36,10 @@ namespace SM_MentalHealthApp.Shared
         public int SmeUserId { get; set; }
         
         /// <summary>
-        /// Billing Account ID: CompanyId if company, SmeUserId if individual
+        /// Billing Account ID (references BillingAccounts table)
         /// </summary>
         [Required]
-        public int BillingAccountId { get; set; }
+        public long BillingAccountId { get; set; }
         
         /// <summary>
         /// Type of billing account: "Company" or "Individual"
@@ -196,7 +196,7 @@ namespace SM_MentalHealthApp.Shared
         public List<SmeInvoiceLineDto> InvoiceLines { get; set; } = new();
         
         // Company billing fields
-        public int? BillingAccountId { get; set; }
+        public long? BillingAccountId { get; set; }
         public string? BillingAccountType { get; set; }
         public string? CompanyName { get; set; }
         public List<string> SmeNames { get; set; } = new(); // All SME names for company invoices
